refactor Dialog 패키지
