+++ feature/board/src/main/java/com/goalpanzi/mission_mate/feature/board/component/dialog/RequestDeleteMissionDialog.kt
@@ -1,9 +1,8 @@
-package com.goalpanzi.mission_mate.feature.board.component
+package com.goalpanzi.mission_mate.feature.board.component.dialog
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.window.DialogProperties
 import com.goalpanzi.mission_mate.core.designsystem.component.MissionMateDialog
 import com.goalpanzi.mission_mate.feature.board.R
