+++ feature/board/src/main/java/com/goalpanzi/mission_mate/feature/board/screen/BoardMissionDetailScreen.kt
@@ -45,7 +45,7 @@ import com.goalpanzi.mission_mate.core.designsystem.theme.MissionMateTypography
 import com.goalpanzi.mission_mate.core.designsystem.theme.component.MissionMateTopAppBar
 import com.goalpanzi.mission_mate.core.designsystem.theme.component.NavigationType
 import com.goalpanzi.mission_mate.feature.board.R
-import com.goalpanzi.mission_mate.feature.board.component.RequestDeleteMissionDialog
+import com.goalpanzi.mission_mate.feature.board.component.dialog.RequestDeleteMissionDialog
 import com.goalpanzi.mission_mate.feature.board.model.uimodel.MissionUiModel
 import com.goalpanzi.mission_mate.feature.onboarding.model.VerificationTimeType
 import com.goalpanzi.mission_mate.feature.onboarding.util.styledTextWithHighlights
